# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Staffio HR Management API
  version: 1.0.0
  description: |
    API do sistema de gestão de RH (Staffio). Backend em Go (Gin).

    Autenticação é baseada em sessão (cookie). A maioria dos endpoints exige
    sessão autenticada e os endpoints de escrita respeitam a hierarquia de papéis:
    - `owner`   (dono)  – acesso total
    - `admin`   (admin) – gerencia equipe, edita candidatos
    - `recruit` (recrutador) – edita candidatos e departamentos, sem relatórios/visão geral
    - `viewer`  (visualizador) – somente leitura; bloqueado para métodos não-GET

    Vários endpoints renderizam páginas HTML (servidor). Apenas os endpoints
    prefixados com `/api/` retornam JSON estruturado.
  contact:
    name: Staffio
servers:
  - url: http://localhost:8000
    description: Servidor local
...
paths:
...
  /login:
    get:
      tags: [Pages, Auth]
      summary: Página de login
      description: Renderiza o formulário de login (HTML).
      responses:
        '200':
          description: Página de login renderizada
          content:
            text/html:
              schema: { type: string }
    post:
      tags: [Auth]
      summary: Autenticar usuário
      description: Valida credenciais e cria uma sessão (cookie). Limitado a 10 requisições por minuto por IP.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  description: Nome de usuário
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '302':
          description: Redireciona para /dashboard em caso de sucesso
        '400':
          description: Campos obrigatórios ausentes
        '401':
          description: Credenciais incorretas
  /register:
    get:
      tags: [Pages, Auth]
      summary: Página de registro
      responses:
        '200':
          description: Página de registro renderizada
          content:
            text/html: { schema: { type: string } }
    post:
      tags: [Auth]
      summary: Registrar nova conta
      description: Cria uma organização e um usuário dono (owner). Limitado a 10 requisições por minuto por IP.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password, email]
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Conta criada, redireciona para a página de login
        '400':
          description: Campo obrigatório ausente ou usuário já existe
        '500':
          description: Erro interno
  /logout:
    get:
      tags: [Auth]
      summary: Encerrar sessão
      description: Destrói a sessão e redireciona para /login. Requer autenticação.
      security:
        - sessionAuth: []
      responses:
        '302':
          description: Redireciona para /login
  /recuperateaccount:
    get:
      tags: [Pages, Auth]
      summary: Página de recuperação de senha
      responses:
        '200':
          description: Página de recuperação renderizada
          content:
            text/html: { schema: { type: string } }
    post:
      tags: [Auth]
      summary: Solicitar recuperação de senha
      description: Envia e-mail com link de redefinição de senha. Limitado a 5 requisições por minuto por IP.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: E-mail de recuperação enviado (resposta idêntica se o e-mail não existir, por segurança)
        '400':
          description: E-mail obrigatório
        '500':
          description: Erro ao enviar e-mail
  /reset-password:
    get:
      tags: [Pages, Auth]
      summary: Página de redefinição de senha
      parameters:
        - name: token
          in: query
          required: true
          description: Token de redefinição recebido por e-mail
          schema: { type: string }
      responses:
        '200':
          description: Página de redefinição renderizada
          content:
            text/html: { schema: { type: string } }
        '400':
          description: Token inválido, ausente ou expirado
    post:
      tags: [Auth]
      summary: Redefinir senha
      description: Define uma nova senha usando o token. Limitado a 5 requisições por minuto por IP.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token, new_password, confirm_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  format: password
                  minLength: 6
                confirm_password:
                  type: string
                  format: password
      responses:
        '302':
          description: Redireciona para /login em caso de sucesso
        '400':
          description: Token inválido/expirado ou senhas não conferem
  /recover:
    get:
      tags: [Pages, Auth]
      summary: Página de recuperação (autenticado)
      description: Requer autenticação. Renderiza a página de recuperação.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Página de recuperação renderizada
          content:
            text/html: { schema: { type: string } }

  # ============================ LANDING / UTIL ============================
...
  /dashboard:
    get:
      tags: [Pages]
      summary: Dashboard
      description: Lista candidatos da organização com busca e paginação (20 por página). Requer autenticação; recrutadores são redirecionados para /employees.
      security:
        - sessionAuth: []
      parameters:
        - name: search
          in: query
          required: false
          description: Filtra por nome ou e-mail
          schema: { type: string }
        - name: all
          in: query
          required: false
          description: "`true` mostra todos (sem limite); `false` limita a 100.000 registros"
          schema: { type: string, enum: ['true', 'false'] }
      responses:
        '200':
          description: Dashboard renderizado
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona para /login se não autenticado

  # ============================ EMPLOYEES ============================
  /employees:
    get:
      tags: [Pages, Employees]
      summary: Listar candidatos (página)
      description: Página com listagem paginada e contagens por status. Requer autenticação.
      security:
        - sessionAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 1, minimum: 1 }
        - name: status
          in: query
          required: false
          description: Filtro de status (interviewing = pending, hired = contractors)
          schema:
            type: string
            enum: [all, interviewing, hired, rejected]
            default: all
      responses:
        '200':
          description: Página de candidatos renderizada
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona para /login se não autenticado
    post:
      tags: [Employees]
      summary: Criar candidato
      description: Cria um novo candidato (status inicial pending). Requer papel com permissão de edição (owner, admin, recruit).
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [full_name, email, position]
              properties:
                full_name:
                  type: string
                email:
                  type: string
                  format: email
                position:
                  type: string
                description:
                  type: string
                photo:
                  type: string
                  format: binary
                  description: Foto opcional (JPG, PNG, GIF, WebP, máx. 5 MB)
      responses:
        '302':
          description: Redireciona para /employees em caso de sucesso
        '400':
          description: Campos obrigatórios ausentes ou upload inválido
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão de edição
        '500':
          description: Erro ao salvar no banco
  /employees/download:
    get:
      tags: [Employees]
      summary: Exportar candidatos em CSV
      description: Baixa a lista de candidatos filtrada em formato CSV. Requer autenticação.
      security:
        - sessionAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, interviewing, hired, rejected]
            default: all
      responses:
        '200':
          description: Arquivo CSV
          content:
            text/csv:
              schema: { type: string }
        '500':
          description: Erro ao exportar
  /employees/{id}/edit:
    get:
      tags: [Pages, Employees]
      summary: Página de edição de candidato
      description: Renderiza o formulário de edição. Requer permissão de edição.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Página de edição renderizada
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona para /employees se não encontrado ou sem permissão
    post:
      tags: [Employees]
      summary: Atualizar candidato
      description: Atualiza dados do candidato e, opcionalmente, a foto. Requer permissão de edição.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [full_name, email, position]
              properties:
                full_name:
                  type: string
                email:
                  type: string
                  format: email
                position:
                  type: string
                description:
                  type: string
                photo:
                  type: string
                  format: binary
      responses:
        '302':
          description: Redireciona para /employees em caso de sucesso
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
  /employees/{id}/status:
    post:
      tags: [Employees]
      summary: Atualizar status do candidato
      description: Altera o status (pending/contractors/rejected) e a data de contratação. Ao rejeitar, registra um desligamento (termination). Recrutadores não podem rejeitar candidatos ligados a membros superiores.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  description: Novo status do candidato
                  enum: [pending, contractors, rejected]
                hire_date:
                  type: string
                  format: date
                  description: Data de contratação (para status contractors)
      responses:
        '302':
          description: Redireciona para /employees em caso de sucesso
        '400':
          description: Status obrigatório
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
  /employees/{id}/absence:
    post:
      tags: [Employees]
      summary: Registrar ausência
      description: Registra uma ausência para o candidato e incrementa o contador. Requer permissão de edição.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Ausência registrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Absence marked
        '400':
          description: Sem permissão
        '401':
          description: Não autenticado
        '404':
          description: Candidato não encontrado
        '500':
          description: Erro ao registrar ausência
  /employees/{id}:
    delete:
      tags: [Employees]
      summary: Excluir candidato (API)
      description: Exclui o candidato e registra um desligamento (termination). Remove a foto armazenada. Recrutadores não podem excluir candidatos ligados a membros superiores.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Candidato excluído
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Employee deleted successfully
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
        '404':
          description: Candidato não encontrado
        '500':
          description: Erro ao excluir
  /employees/{id}/delete:
    post:
      tags: [Employees]
      summary: Excluir candidato (formulário)
      description: Exclui o candidato via formulário (HTML) e redireciona. Registra desligamento e remove foto. Requer permissão de edição.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '302':
          description: Redireciona para /employees em caso de sucesso
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
        '500':
          description: Erro ao excluir
  /badge/{id}:
    get:
      tags: [Pages, Employees]
      summary: Crachá do candidato
      description: Renderiza o crachá (ID card) do candidato. Requer autenticação.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Crachá renderizado
          content:
            text/html: { schema: { type: string } }
        '404':
          description: Candidato não encontrado

  # ============================ DEPARTMENTS ============================
  /department:
    get:
      tags: [Pages, Departments]
      summary: Página de departamentos
      description: Lista departamentos e funcionários. Recrutadores são redirecionados para /employees.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Página de departamentos renderizada
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona se não autenticado ou se é recrutador
        '500':
          description: Erro ao buscar dados
    post:
      tags: [Departments]
      summary: Criar departamento
      description: Cria um novo departamento. Requer permissão de gestão de equipe (owner, admin).
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [code, name]
              properties:
                code:
                  type: string
                name:
                  type: string
                boss_id:
                  type: integer
                  description: ID do gerente (opcional)
      responses:
        '302':
          description: Redireciona para /department em caso de sucesso
        '400':
          description: Campos obrigatórios ausentes ou ID de gerente inválido
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
        '500':
          description: Erro ao salvar departamento
  /department/{id}:
    get:
      tags: [Pages, Departments]
      summary: Detalhe do departamento
      description: Mostra o departamento, seus membros e todos os funcionários disponíveis. Recrutadores são redirecionados.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Página do departamento renderizada
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona se não autenticado ou recrutador
        '404':
          description: Departamento não encontrado
  /department/{id}/add_employee:
    post:
      tags: [Departments]
      summary: Adicionar funcionário ao departamento
      description: Atribui um funcionário ao departamento. Requer permissão de atribuição (owner, admin, recruit).
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [employee_id]
              properties:
                employee_id:
                  type: integer
      responses:
        '302':
          description: Redireciona para /department/{id}
        '400':
          description: employee_id obrigatório
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
        '500':
          description: Erro ao atribuir
  /department/{id}/remove_employee:
    post:
      tags: [Departments]
      summary: Remover funcionário do departamento
      description: Desvincula um funcionário do departamento. Requer permissão de atribuição.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [employee_id]
              properties:
                employee_id:
                  type: integer
      responses:
        '302':
          description: Redireciona para /department/{id}
        '400':
          description: employee_id obrigatório
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
        '500':
          description: Erro ao remover
  /department/{id}/delete:
    post:
      tags: [Departments]
      summary: Excluir departamento
      description: Desvincula os funcionários e exclui o departamento. Requer gestão de equipe (owner, admin).
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '302':
          description: Redireciona para /department em caso de sucesso
        '401':
          description: Não autenticado
        '403':
          description: Sem permissão
        '500':
          description: Erro ao excluir

  # ============================ OVERVIEW ============================
  /overview:
    get:
      tags: [Pages, Overview]
      summary: Visão geral
      description: Página de visão geral. Recrutadores são redirecionados para /employees.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Página de visão geral renderizada
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona se não autenticado ou recrutador
  /api/overview/departments:
    get:
      tags: [API, Overview]
      summary: Contagem de funcionários por departamento
      description: Retorna a contagem de funcionários agrupada por departamento. Requer autenticação.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Dados por departamento
          content:
            application/json:
              schema:
                type: object
                properties:
                  departments:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        count: { type: integer }
        '401':
          description: Não autenticado
  /api/overview/employees:
    get:
      tags: [API, Overview]
      summary: Série de candidatos e desligados por dia
      description: Retorna candidatos criados por dia e desligados/rejeitados por dia. Requer autenticação.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Séries temporais
          content:
            application/json:
              schema:
                type: object
                properties:
                  employees:
                    type: array
                    items:
                      type: object
                      properties:
                        day: { type: string }
                        count: { type: integer }
                  fired:
                    type: array
                    items:
                      type: object
                      properties:
                        day: { type: string }
                        count: { type: integer }
        '401':
          description: Não autenticado
        '500':
          description: Erro na consulta

  # ============================ REPORTS ============================
  /report:
    get:
      tags: [Pages, Reports]
      summary: Página de relatórios
      description: Lista relatórios com totais de ausências, contratações e desligamentos. Recrutadores são redirecionados.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Página de relatórios renderizada
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona se não autenticado ou recrutador
  /report/new:
    get:
      tags: [Pages, Reports]
      summary: Página de novo relatório
      responses:
        '200':
          description: Página de novo relatório renderizada
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona se não autenticado ou recrutador
    post:
      tags: [Reports]
      summary: Criar relatório
      description: Cria um relatório de ausências, contratações ou desligamentos para um período. Recrutadores são redirecionados.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                report-type:
                  type: string
                  enum: [absences, hired, fired]
                  default: absences
                time-report:
                  type: string
                  enum: [week, month, year]
                  default: week
                  description: Período (week=7 dias, month=30, year=365)
      responses:
        '302':
          description: Redireciona para /report em caso de sucesso
        '302':
          description: Redireciona se não autenticado ou recrutador
  /report/{id}:
    get:
      tags: [Pages, Reports]
      summary: Detalhe do relatório
      description: Mostra o relatório com resumo textual e estatísticas. Recrutadores são redirecionados.
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Página do relatório renderizada
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona se não autenticado, recrutador ou relatório não encontrado
...
  /config:
    get:
      tags: [Pages, Config]
      summary: Página de configuração
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Página de configuração renderizada
          content:
            text/html: { schema: { type: string } }
        '302':
          description: Redireciona para /login se não autenticado
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie de sessão definido pelo servidor após login
...